requirements.txt
boto3>=1.26.0
azure-mgmt-compute>=29.1.0
azure-mgmt-monitor>=5.0.0
azure-identity>=1.12.0
google-cloud-compute>=1.11.0
google-cloud-billing>=1.10.0
pandas>=2.0.0
python-dotenv>=1.0.0
schedule>=1.2.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
